// server/src/types/enums/status-code.ts
export enum STATUS_CODE {
  SUCCESS_OK = 200,
  CREATED = 201,
  ACCEPTED = 203,
  PERMANENT_REDIRECT = 301,
  TEMPORARY_REDIRECT = 302,
  NOT_MODIFIED = 304,
  BAD_REQUEST = 400,
  UNAUTHORIZED_ERROR = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  INVALID_TOKEN = 498,
  INTERNAL_SERVER_ERROR = 500,
}